Separate e2e.go out into multiple files
